requirements.txt
Package                   Version
------------------------- -----------
altair                    5.4.1
attrs                     24.2.0
blinker                   1.8.2
cachetools                5.5.0
certifi                   2024.8.30
charset-normalizer        3.4.0
click                     8.1.7
contourpy                 1.1.1
cycler                    0.12.1
filelock                  3.16.1
fonttools                 4.55.0
fsspec                    2024.10.0
gitdb                     4.0.11
GitPython                 3.1.43
idna                      3.10
importlib_resources       6.4.5
Jinja2                    3.1.4
joblib                    1.4.2
jsonschema                4.23.0
jsonschema-specifications 2023.12.1
kiwisolver                1.4.7
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.7.5
mdurl                     0.1.2
mpmath                    1.3.0
narwhals                  1.13.5
networkx                  3.1
numpy                     1.24.4
nvidia-cublas-cu12        12.1.3.1
nvidia-cuda-cupti-cu12    12.1.105
nvidia-cuda-nvrtc-cu12    12.1.105
nvidia-cuda-runtime-cu12  12.1.105
nvidia-cudnn-cu12         9.1.0.70
nvidia-cufft-cu12         11.0.2.54
nvidia-curand-cu12        10.3.2.106
nvidia-cusolver-cu12      11.4.5.107
nvidia-cusparse-cu12      12.1.0.106
nvidia-nccl-cu12          2.20.5
nvidia-nvjitlink-cu12     12.6.77
nvidia-nvtx-cu12          12.1.105
packaging                 24.2
pandas                    2.0.3
pillow                    10.4.0
pip                       24.3.1
pkgutil_resolve_name      1.3.10
protobuf                  5.28.3
pyarrow                   17.0.0
pydeck                    0.9.1
Pygments                  2.18.0
pyparsing                 3.1.4
python-dateutil           2.9.0.post0
pytz                      2024.2
referencing               0.35.1
requests                  2.32.3
rich                      13.9.4
rpds-py                   0.20.1
scikit-learn              1.3.2
scipy                     1.10.1
seaborn                   0.13.2
setuptools                75.3.0
six                       1.16.0
smmap                     5.0.1
streamlit                 1.40.1
sympy                     1.13.3
tenacity                  9.0.0
threadpoolctl             3.5.0
toml                      0.10.2
torch                     2.4.1
torchvision               0.19.1
tornado                   6.4.1
triton                    3.0.0
typing_extensions         4.12.2
tzdata                    2024.2
urllib3                   2.2.3
watchdog                  4.0.2
wheel                     0.45.0
zipp                      3.20.2
